Change parent_id type to unsignedBigInteger
Change parent_id column type to unsignedBigInteger since laravel >=5.8 use bigIncrements as the default type

+++ src/NestedSet.php
@@ -40,7 +40,7 @@ class NestedSet
     {
         $table->unsignedInteger(self::LFT)->default(0);
         $table->unsignedInteger(self::RGT)->default(0);
-        $table->unsignedInteger(self::PARENT_ID)->nullable();
+        $table->unsignedBigInteger(self::PARENT_ID)->nullable();
 
         $table->index(static::getDefaultColumns());
     }
